+++ client/kit/src/routes/(authenticated)/+page.svelte
